requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
alpha-vantage==2.3.1
amqp==5.1.1
appdirs==1.4.4
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beartype==0.12.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
Django==4.1.4
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
finnhub-python==2.4.16
flet==0.3.2
frozendict==2.3.4
frozenlist==1.3.3
github==1.2.7
h11==0.14.0
html5lib==1.1
idna==3.4
kombu==5.2.4
lxml==4.9.2
multidict==6.0.4
multitasking==0.0.11
newsapi-python==0.2.6
numpy==1.24.1
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pgeocode==0.4.0
prompt-toolkit==3.0.38
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
repath==0.9.0
requests==2.28.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
undetected-chromedriver==3.4.6
urllib3==1.26.14
vine==5.0.0
watchdog==2.2.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
yfinance==0.2.9
dj_database_url== 1.2.0
gunicorn
psycopg2-binary